% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticMK.R
\name{asymptoticMK}
\alias{asymptoticMK}
\title{Asymtotic MKT developed in Messer, P. W. & Petrov, D. A. Frequent adaptation and the McDonald-Kreitman test. Proceedings of the National Academy of Sciences 110, 8615–8620 (2013).
Directly retrieved from: https://github.com/MesserLab/asymptoticMK
Date = 30/11/2016
Author = Sergi Hervás, Marta Coronado}
\usage{
asymptoticMK(daf, divergence, xlow, xhigh)
}
\arguments{
\item{daf}{dad file}

\item{divergence}{div file}

\item{xlow}{fit curve}

\item{xhigh}{fit curv}
}
\value{
None
}
\description{
The standard McDonald and Kreitman test (MKT) is used to detect the signature of selection at the molecular level. The MKT compares the amount of variation within a species (polymorphism, P) to the divergence (D) between species at two types of sites, one of which is putatively netral and used as the reference to detect selection at the other type of site. In the standard MKT, these sites are synonymous (putatively neutral, 0) and non-synonymous sites (selected sites, i) in a coding region. Under strict neutrality, the ratio of the number of selected and neutral polymorphic sites (Pi/P0) is equal to the ratio of the number of selected and neutral divergence sites (Di/D0).
}
