% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{check_input}
\usage{
check_input(daf, divergence, xlow, xhigh)
}
\arguments{
\item{daf}{daf file}

\item{divergence}{divergence file}

\item{xlow}{trimming values below this daf threshold}

\item{xhigh}{trimming values above this daf threshold}
}
\value{
None
}
\description{
Error handling. Checks input data and returns errors whan it is malformed.
Author = Sergi Hervas, Jesús Murga
}
