% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergenceParameters.R
\name{divergenceParameters}
\alias{divergenceParameters}
\title{KaKs}
\usage{
divergenceParameters(daf = "Data frame containing the DAF, Pi and P0",
  divergence = "Data frame that contains sites analyzed and divergencen 0fold and 4fold")
}
\arguments{
\item{daf}{daf file}

\item{divergence}{divergence file}
}
\value{
Standard McDonald and Kreitman Test
}
\description{
KaKs
}
\details{
KaKs
}
\examples{
## Load your Derived Allele Frequency and Divergence files
daf <- mydafdata
div <- mydivergencedata
## Run the function
# divergenceParameters(daf, div)

}
