% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMK.R
\name{iMK}
\alias{iMK}
\title{iMK}
\usage{
iMK(daf, divergence, xlow, xhigh, seed, plot = FALSE)
}
\arguments{
\item{daf}{data frame containing DAF, Pi and P0 values}

\item{divergence}{data frame containing divergent and analyzed sites for selected (i) and neutral (0) classes}

\item{xlow}{lower limit for asymptotic alpha fit}

\item{xhigh}{higher limit for asymptotic alpha fit}

\item{seed}{seed value (optional). No seed by default}

\item{plot}{report plots of daf, alpha and negative selection fractions (optional). Default is FALSE}
}
\value{
iMK
}
\description{
iMK: alpha asymptotic + negative selection (d,b,f)

#details details here
}
\examples{
## Without plot
iMK(mydafdata, mydivergencedata, 0, 0.9)
## With plot
iMK(mydafdata, mydivergencedata, 0, 0.9, plot=TRUE)

}
