% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrativeMKT.R
\name{integrativeMKT}
\alias{integrativeMKT}
\title{integrativeMKT}
\usage{
integrativeMKT(daf, divergence, xlow, xhigh, seed)
}
\arguments{
\item{daf}{daf file}

\item{divergence}{div file}

\item{xlow}{fit curve}

\item{xhigh}{fit curv}

\item{seed}{seed value (optional). No seed by default}
}
\value{
Execute all the MKT extensions
}
\description{
\code{itnegrativeMKT()} put details here
}
\details{
put description here
}
\examples{
## Load your Derived Allele Frequency and Divergence files
daf <- mydafdata
div <- mydivergencedata
## Run the function
integrativeMKT(daf, div, 0, 0.9)

}
