% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPopHuman.R
\name{loadPopHuman}
\alias{loadPopHuman}
\title{Load PopHuman dataset}
\usage{
loadPopHuman()
}
\value{
PopHumanData object loaded in the workspace
}
\description{
Load PopHuman dataset with information regarding protein coding gene annotations
}
\details{
This function loads PopHuman data (Casillas et al. 2018 Nucleic Acids Research, http://pophuman.uab.cat/) into the current workspace. Data is stored in a dataframe named PopHumanData, which includes population genetics estimates (nucleotide diversity, divergence, basic tests of neutrality, recombination rates, etc.) regarding each protein coding gene for 26 worldwide Homo sapiens populations from the 1000 Genomes Project (The 1000 Genomes Project Consortium 2012 Nature, The 1000 Genomes Project Consortium 2015 Nature).
}
\examples{
## Load PopHuman data if necessary. This process may take several seconds to complete.
# loadPopHuman()

}
\keyword{PopData}
