% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/mydivergencedata.R
\docType{data}
\name{mydafdata}
\alias{mydafdata}
\alias{mydafdata}
\title{Example data frames}
\format{An object of class \code{data.frame} with 20 rows and 3 columns.}
\usage{
mydafdata

mydafdata
}
\description{
Containing basic information to execute the funcionts

Containing basic information to execute the funcionts
}
\keyword{datasets}
