% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydivergencedata.R
\docType{data}
\name{mydivergencedata}
\alias{mydivergencedata}
\title{Example data frames}
\format{An object of class \code{data.frame} with 1 rows and 6 columns.}
\usage{
mydivergencedata
}
\description{
Containing basic information to execute the funcionts
}
\keyword{datasets}
