% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotALPHA.R
\name{plotALPHA}
\alias{plotALPHA}
\title{Alpha plot 
Date = 30/11/2016
Author = Marta Coronado, Sergi Hervas}
\usage{
plotALPHA(x, y, df, xlow, xhigh)
}
\arguments{
\item{x}{daf file}

\item{y}{divergence file}

\item{df}{alpha asymptotic output (iMK$asymptoticMK)}

\item{xlow}{fit curve}

\item{xhigh}{fit curve}
}
\value{
None
}
\description{
Alpha plot 
Date = 30/11/2016
Author = Marta Coronado, Sergi Hervas
}
