% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDAF.R
\name{plotDAF}
\alias{plotDAF}
\title{Daf plot 
Date = 30/11/2016
Author = Marta Coronado, Sergi Hervàs}
\usage{
plotDAF(x)
}
\arguments{
\item{x}{daf file: daf, pN, pS}
}
\value{
None
}
\description{
Daf plot 
Date = 30/11/2016
Author = Marta Coronado, Sergi Hervàs
}
