% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIMK.R
\name{plotIMK}
\alias{plotIMK}
\title{iMK horizontal plot 
Date = 30/11/2016
Author = Marta Coronado, Sergi Hervàs}
\usage{
plotIMK(df)
}
\arguments{
\item{df}{table from iMK (1 observation) = iMK$iMK}
}
\value{
None
}
\description{
iMK horizontal plot 
Date = 30/11/2016
Author = Marta Coronado, Sergi Hervàs
}
