% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.R
\name{standard}
\alias{standard}
\title{mkt_standard}
\usage{
standard(daf = "Data frame containing the DAF, Pn and Ps",
  divergence = "Data frame that contains sites analyzed and divergencen 0fold and 4fold")
}
\arguments{
\item{daf}{daf file}

\item{divergence}{divergence file}
}
\value{
Standard McDonald and Kreitman Test
}
\description{
The standard McDonald and Kreitman test (MKT) is used to detect the signature of selection at the molecular level. The MKT compares the amount of variation within a species (polymorphism, P) to the divergence (D) between species at two types of sites, one of which is putatively netral and used as the reference to detect selection at the other type of site. In the standard MKT, these sites are synonymous (putatively neutral, 0) and non-synonymous sites (selected sites, i) in a coding region. Under strict neutrality, the ratio of the number of selected and neutral polymorphic sites (Pi/P0) is equal to the ratio of the number of selected and neutral divergence sites (Di/D0).
}
\details{
\code{mkt_standard()} MKT standard formula
Load your Derived Allele Frequency file (remember you can use 10 or 20 categories) and Divergence file (it contains de divergents and analysed sites in the synonimous and non synonimous categories)
}
\examples{
#Load your Derived Allele Frequency file and Divergence file
daf<-read.table("/home/jmurga/MKT/data.daf.txt",header=TRUE)
div<-read.table("/home/jmurga/MKT/data.divergence.txt",header=TRUE)
#Run the function!
standard(daf,div)
}
