% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{MKT corrected with DGRP method}
\usage{
theme_Publication(base_size = 14, base_family = "helvetica")
}
\arguments{
\item{base_size}{base size required from theme_Publication}

\item{base_family}{font to load in theme_Publication}
}
\value{
MKT corrected by the DGRP method
}
\description{
Date = 30/11/2016
Author = Sergi Hervás, Marta Coronado
}
\details{
Default theme use for plot images. It has enough quality to publish anything
}
