% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themePublication.R
\name{themePublication}
\alias{themePublication}
\title{ggplot theme for publication ready Plots}
\usage{
themePublication(base_size = 14, base_family = "sans")
}
\arguments{
\item{base_size}{base size required from themePublication}

\item{base_family}{font to load in themePublication}
}
\value{
plot theme
}
\description{
Date = 04/07/2015
Author = Koundinya Desiraju
}
\details{
Default theme used for plot images. From http://rpubs.com/Koundy/71792
}
\examples{
# themePublication(14, "sans")

}
