% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticMK.R
\name{asymptoticMK}
\alias{asymptoticMK}
\title{asymptoticMK}
\usage{
asymptoticMK(daf, divergence, xlow, xhigh, seed)
}
\arguments{
\item{daf}{daf file}

\item{divergence}{div file}

\item{xlow}{fit curve}

\item{xhigh}{fit curv}

\item{seed}{seed value (optional). No seed by default}
}
\value{
None
}
\description{
\code{asymptoticMK} developed in Messer, P. W. & Petrov, D. A. Frequent adaptation and the McDonald-Kreitman test. Proceedings of the National Academy of Sciences 110, 8615–8620 (2013).
Directly retrieved from: https://github.com/MesserLab/asymptoticMK
Load your Derived Allele Frequency file (remember you can use 10 or 20 categories) and Divergence file (it contains de divergents and analysed sites in the synonimous and non synonimous categories)
}
\details{
The standard McDonald and Kreitman test (MKT) is used to detect the signature of selection at the molecular level. The MKT compares the amount of variation within a species (polymorphism, P) to the divergence (D) between species at two types of sites, one of which is putatively netral and used as the reference to detect selection at the other type of site. In the standard MKT, these sites are synonymous (putatively neutral, 0) and non-synonymous sites (selected sites, i) in a coding region. Under strict neutrality, the ratio of the number of selected and neutral polymorphic sites (Pi/P0) is equal to the ratio of the number of selected and neutral divergence sites (Di/D0).
}
\examples{
daf<-read.table("/home/jmurga/MKT/Test/data.daf.txt",header=TRUE)
div<-read.table("/home/jmurga/MKT/Test/data.divergence.txt",header=TRUE)
asymptoticMK(daf=daf,divergence=div,xlow=0,xhigh=1)

}
