% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticMK.R
\name{asymptoticMK}
\alias{asymptoticMK}
\title{asymptoticMK}
\usage{
asymptoticMK(daf, divergence, xlow, xhigh, seed)
}
\arguments{
\item{daf}{data frame containing DAF, Pi and P0 values}

\item{divergence}{data frame containing divergent and analyzed sites for selected (i) and neutral (0) classes}

\item{xlow}{trimming values below this daf threshold}

\item{xhigh}{trimming values above this daf threshold}

\item{seed}{seed value (optional). No seed by default}
}
\value{
Estimation of alpha asymptotic value and details of the model fit
}
\description{
\code{asymptoticMK} developed in "Haller BC, Messer PW. asymptoticMK: A Web-Based Tool for the Asymptotic McDonald-Kreitman Test. G3 (Bethesda). 2017 May 5;7(5):1569-1575".
Adapted from: http://github.com/MesserLab/asymptoticMK
}
\details{
The standard McDonald and Kreitman test (MKT) is used to detect the signature of selection at the molecular level. The MKT compares the amount of variation within a species (polymorphism, P) to the divergence (D) between species at two types of sites, one of which is putatively netral and used as the reference to detect selection at the other type of site. In the standard MKT, these sites are synonymous (putatively neutral, 0) and non-synonymous sites (selected sites, i) in a coding region. Under strict neutrality, the ratio of the number of selected and neutral polymorphic sites (Pi/P0) is equal to the ratio of the number of selected and neutral divergence sites (Di/D0).
}
\examples{
# asymptoticMK(mydafdata, mydivergencedata, 0, 0.9)

}
