% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{check_input}
\usage{
check_input(daf, divergence, xlow, xhigh)
}
\arguments{
\item{daf}{daf file}

\item{divergence}{divergence file}

\item{xlow}{fit curve}

\item{xhigh}{fit curv}
}
\value{
None
}
\description{
Error handling. Checks input data for asymptoticMKAuthor = Sergi Hervas, Jesús Murga
}
