% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Error handling to compute correctly each MKT type. ¿mix between Messer & Haller? Preguntar Sergi
Date = 30/11/2016
' Author = Sergi Hervás, Jesús Murga}
\usage{
check_input(daf, divergence, xlow, xhigh)
}
\arguments{
\item{daf}{dad file}

\item{divergence}{divergence file}

\item{xlow}{fit curve}

\item{xhigh}{fit curv}
}
\value{
None
}
\description{
Error handling to compute correctly each MKT type. ¿mix between Messer & Haller? Preguntar Sergi
Date = 30/11/2016
' Author = Sergi Hervás, Jesús Murga
}
