% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeMKT.R
\name{completeMK}
\alias{completeMK}
\title{Complete MK methodologies}
\usage{
completeMK(daf, divergence, xlow, xhigh, seed)
}
\arguments{
\item{daf}{data frame containing DAF, Pi and P0 values}

\item{divergence}{data frame containing divergent and analyzed sites for selected (i) and neutral (0) classes}

\item{xlow}{lower limit for asymptotic alpha fit}

\item{xhigh}{higher limit for asymptotic alpha fit}

\item{seed}{seed value (optional). No seed by default}
}
\value{
List with the diverse MKT results: standardMK, FWW, DGRP, asymptoticMK, iMK
}
\description{
MKT calculation using all methodologies included in the package: standardMK, FWW, DGRP, asymptoticMK, iMK.
}
\details{
Perform all MKT derived methodologies at once using the same input data and parameters.
}
\examples{
completeMK(myDafData, myDivergenceData, xlow=0, xhigh=0.9)

}
\keyword{MKT}
