% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeMKT.R
\name{completeMKT}
\alias{completeMKT}
\title{completeMKT}
\usage{
completeMKT(daf, divergence, xlow, xhigh, seed)
}
\arguments{
\item{daf}{data frame containing DAF, Pi and P0 values}

\item{divergence}{data frame containing divergent and analyzed sites for selected (i) and neutral (0) classes}

\item{xlow}{lower limit for asymptotic alpha fit}

\item{xhigh}{higher limit for asymptotic alpha fit}

\item{seed}{seed value (optional). No seed by default}
}
\value{
Execute all the MKT extensions
}
\description{
\code{completeMKT()} put details here
}
\details{
put description here
}
\examples{
# completeMKT(mydafdata, mydivergencedata, 0, 0.9)

}
