% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeMKT.R
\name{completeMKT}
\alias{completeMKT}
\title{completeMKT}
\usage{
completeMKT(daf, divergence, xlow, xhigh, seed)
}
\arguments{
\item{daf}{daf file}

\item{divergence}{div file}

\item{xlow}{trimming values below this daf threshold}

\item{xhigh}{trimming values above this daf threshold}

\item{seed}{seed value (optional). No seed by default}
}
\value{
Execute all the MKT extensions
}
\description{
\code{completeMKT()} put details here
}
\details{
put description here
}
\examples{
## Load your Derived Allele Frequency and Divergence files
daf <- mydafdata
div <- mydivergencedata
## Run the function
# completeMKT(daf, div, 0, 0.9)

}
