% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMKmodel.R
\name{fitMKmodel}
\alias{fitMKmodel}
\title{fitMKmodel}
\usage{
fitMKmodel(alpha_trimmed, f_trimmed, res)
}
\arguments{
\item{alpha_trimmed}{correspond to the vectors with alpha and frequency (DAF) values}

\item{f_trimmed}{correspond to the vectors with alpha and frequency (DAF) values}

\item{res}{ASK SERGI}
}
\value{
None
}
\description{
Core code for two-step nls2() model fit at a given level of precision (res, usually set at 10)
Directly retrieved from: https://github.com/MesserLab/asymptoticMK
}
\details{
Put details here
}
