% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FWW.R
\name{FWW}
\alias{FWW}
\title{mkt_fww}
\usage{
FWW(daf = "Data frame containing the DAF, Pn and Ps",
  divergence = "Data frame that contains sites analyzed and divergencen 0fold and 4fold")
}
\arguments{
\item{daf}{daf file}

\item{divergence}{divergence file}
}
\value{
MKT corrected by the FWW method
}
\description{
\code{mktDGRP()} Perform MKT corrected with FWW method  

Load your Derived Allele Frequency file (remember you can use 10 or 20 categories) and Divergence file (it contains de divergents and analysed sites in the synonimous and non synonimous categories)
}
\details{
The standard McDonald and Kreitman test (MKT) is used to detect the signature of selection at the molecular level. The MKT compares the amount of variation within a species (polymorphism, P) to the divergence (D) between species at two types of sites, one of which is putatively netral and used as the reference to detect selection at the other type of site. In the standard MKT, these sites are synonymous (putatively neutral, 0) and non-synonymous sites (selected sites, i) in a coding region. Under strict neutrality, the ratio of the number of selected and neutral polymorphic sites (Pi/P0) is equal to the ratio of the number of selected and neutral divergence sites (Di/D0). The null hypothesis of neutrality is rejected in a MKT when Di/D0 > Pi/P0. The excess of divergence relative to polymorphism for class i, is interpreted as adaptive selection for a subset of sites i. The fraction of adaptive fixations, alpha.symbol, is estimated from 1-(PN/PS)(Ds/Dn). The significance of the test can be assesed with a Fisher exact test.The estimate of alpha.symbol can be easily biased by the segregation of slightly deleterious non-synonymous substitutions. Specifically, slightly deleterious mutations tend to contribute more to polymorphism than to divergence, and thus, lead to an underestimation of alpha. Bevause they tend to segregate at lower frequencies than do neutral mutations, they can be apartially controled for by removing low frequency polymorphisms from the analysis (Fay et al. 2001). This is known as the FWW method.
}
\examples{
#Load your Derived Allele Frequency file and Divergence file
daf<-read.table("/home/jmurga/MKT/data.daf.txt",header=TRUE)
div<-read.table("/home/jmurga/MKT/data.divergence.txt",header=TRUE)
#Run the function!
FWW(daf,div)

}
