% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMK.R
\name{iMK}
\alias{iMK}
\title{iMK compute all the MKT extensions 
Date = 30/11/2016
Author = Sergi Hervas, Marta Coronado}
\usage{
iMK(daf, divergence, xlow, xhigh, seed)
}
\arguments{
\item{daf}{dad file}

\item{divergence}{div file}

\item{xlow}{fit curve}

\item{xhigh}{curv}

\item{seed}{seed value (optional). No seed by default}
}
\value{
None
}
\description{
EXPLANATION
}
