% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMK.R
\name{iMK}
\alias{iMK}
\title{iMK compute all the MKT extensions 
Date = 30/11/2016
Author = Sergi Hervas, Marta Coronado}
\usage{
iMK(daf, divergence, xlow, xhigh, seed)
}
\arguments{
\item{daf}{dad file}

\item{divergence}{div file}

\item{xlow}{fit curve}

\item{xhigh}{curv}

\item{seed}{seed value (optional). No seed by default}
}
\value{
None
}
\description{
EXPLANATION
}
\examples{
#Load your Derived Allele Frequency file and Divergence file
daf<-read.table("/home/jmurga/MKT/Test/data.daf.txt",header=TRUE)
div<-read.table("/home/jmurga/MKT/Test/data.divergence.txt",header=TRUE)
#Run the function!
iMK(daf,div,0,1)

}
