% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMK.R
\name{iMK}
\alias{iMK}
\title{iMK}
\usage{
iMK(daf, divergence, xlow, xhigh, seed)
}
\arguments{
\item{daf}{daf file}

\item{divergence}{div file}

\item{xlow}{fit curve}

\item{xhigh}{curv}

\item{seed}{seed value (optional). No seed by default}
}
\value{
None
}
\description{
\code{iMK()} iMK method
}
\details{
put some details here
}
\examples{
## Load your Derived Allele Frequency and Divergence files
daf <- mydafdata
div <- mydivergencedata
## Run the function
#iMK(daf, div, 0, 0.9)

}
