% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMK.R
\name{iMK}
\alias{iMK}
\title{iMK compute all the MKT extensions 
Date = 30/11/2016
Author = Sergi Hervás, Marta Coronado}
\usage{
iMK(daf, divergence, xlow, xhigh)
}
\arguments{
\item{daf}{dad file}

\item{divergence}{div file}

\item{xlow}{fit curve}

\item{xhigh}{curv}
}
\value{
None
}
\description{
EXPLANATION
}
