% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMK.R
\name{iMK}
\alias{iMK}
\title{integrative MKT method}
\usage{
iMK(daf, divergence, xlow, xhigh, seed, plot = FALSE)
}
\arguments{
\item{daf}{data frame containing DAF, Pi and P0 values}

\item{divergence}{data frame containing divergent and analyzed sites for selected (i) and neutral (0) classes}

\item{xlow}{lower limit for asymptotic alpha fit}

\item{xhigh}{higher limit for asymptotic alpha fit}

\item{seed}{seed value (optional). No seed by default}

\item{plot}{report plots of daf, alpha and negative selection fractions (optional). Default is FALSE}
}
\value{
iMKT method. List with asymptotic MK table and values, fractions of sites and graphs of DAF, asymptotic alpha model and negative selection fractions (optional).
}
\description{
iMK: MKT using asymptoticMK method and estimation of negative selection fractions (d, b, f)
}
\details{
The integrative MKT (iMKT) allows the estimation of the rate of adaptive evolution (alpha) and the diverse negative selection regimens. iMKT uses asymptotic MK method (Messer and Petrov 2012 PNAS; Haller and Messer 2017 G3) to estimate alpha and the diverse negative selection fractions (d: strongly deleterious, b: weakly deleterious, f: neutral), based on the assumption that weakly deleterious mutations usually do not reach high allele frequencies and therefore, produce the underestimation of alpha at low DAF categories. The fraction of strongly deleterious mutations is estimated as the difference between neutral (0) and selected (i) polymorphic sites relative to the number of analyzed sites: d = 1 - (P0/m0 / Pi/mi). The fraction of weakly deleterious sites (b) corresponds to the relative proportion of selected polymorphic sites that cause the underestimation of alpha at low DAF categories. Finally, the fraction of neutral sites (f) is estimated as: f = 1 - d - b.
}
\examples{
## Without plot
iMK(myDafData, myDivergenceData, xlow=0, xhigh=0.9)
## With plot
iMK(myDafData, myDivergenceData, xlow=0, xhigh=0.9, plot=TRUE)

}
\keyword{MKT}
