% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrativeMKT.R
\name{integrativeMKT}
\alias{integrativeMKT}
\title{integrativeMKT}
\usage{
integrativeMKT(daf, divergence, xlow, xhigh, seed)
}
\arguments{
\item{daf}{daf file}

\item{divergence}{div file}

\item{xlow}{fit curve}

\item{xhigh}{fit curv}

\item{seed}{seed value (optional). No seed by default}
}
\value{
Execute all the MKT extensions
}
\description{
\code{itnegrativeMKT()} put details here
}
\details{
put description here
}
\examples{
#Load your Derived Allele Frequency file and Divergence file
#daf<-read.table("/home/jmurga/MKT/Test/data.daf.txt",header=TRUE)
#div<-read.table("/home/jmurga/MKT/Test/data.divergence.txt",header=TRUE)
#Run the function!
#integrativeMKT(daf,div,0,1)

}
