% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPopFly.R
\name{loadPopFly}
\alias{loadPopFly}
\title{loadPopFly}
\usage{
loadPopFly()
}
\value{
None
}
\description{
Load PopFly dataset
}
\details{
This function loads PopFly data into the current workspace. Data is stored in a dataframe named PopFlyData.
}
\examples{
loadPopFly()

}
