% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPopFly.R
\name{loadPopFly}
\alias{loadPopFly}
\title{Load PopFly dataset}
\usage{
loadPopFly()
}
\value{
PopFlyData object loaded in the workspace
}
\description{
Load PopFly dataset with information regarding protein coding gene annotations
}
\details{
This function loads PopFly data (Hervas et al. 2017 Bioinformatics, http://popfly.uab.cat/) into the current workspace. Data is stored in a dataframe named PopFlyData, which includes population genetics estimates (nucleotide diversity, divergence, basic tests of neutrality, recombination rates, etc.) regarding each protein coding gene for 16 worldwide wild-derived Drosophila melanogaster populations from the Drosophila Genome Nexus project (Lack et al. 2015 Genetics, Lack et al. 2016 MBE).
}
\examples{
## Load PopFly data if necessary. This process may take several seconds to complete.
# loadPopFly()

}
\keyword{PopData}
