% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPopHuman.R
\name{loadPopHuman}
\alias{loadPopHuman}
\title{loadPopHuman}
\usage{
loadPopHuman()
}
\value{
None
}
\description{
Load PopHuman dataset
}
\details{
This function loads PopHuman data (http://pophuman.uab.cat/) into the current workspace. Data is stored in a dataframe named PopFlyData.
}
\examples{
loadPopFly()

}
