% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_datasets.R
\name{multiple_datasets}
\alias{multiple_datasets}
\title{Execute any test  with all the files in a directory. The files need to be called id.daf.txt and id.divergence.txt
Date = 30/11/2016
Author = Jesús Murga, Marta Coronado}
\usage{
multiple_datasets(directory = "Directory", test = "MKT type")
}
\arguments{
\item{directory}{dad file}

\item{test}{divergence file}
}
\value{
None
}
\description{
Execute any test  with all the files in a directory. The files need to be called id.daf.txt and id.divergence.txt
Date = 30/11/2016
Author = Jesús Murga, Marta Coronado
}
