% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleDatasets.R
\name{multipleDatasets}
\alias{multipleDatasets}
\title{multipleDatasets}
\usage{
multipleDatasets(directory = "Directory", test = c("DGRP", "FWW", "ALL"),
  fullanalyis = TRUE/FALSE, idlist = "listfile")
}
\arguments{
\item{directory}{dad file}

\item{test}{divergence file}

\item{fullanalyis}{TRUE indicates execute the analysis for all the files in a folder}

\item{idlist}{List of genes or datasets to analyze}
}
\value{
None
}
\description{
\code{multipleDatsets()} Execute any test  with all the files in a directory. The files need to be called id.daf.txt and id.divergence.txt
}
\details{
put a description here
}
