% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleDatasets.R
\name{multipleDatasets}
\alias{multipleDatasets}
\title{multipleDatasets}
\usage{
multipleDatasets(directory = directory, test = c("standard", "DGRP", "FWW",
  "asymptotic", "iMK"), xlow = 0, xhigh = 1, fullAnalysis = TRUE/FALSE,
  idList = "NA")
}
\arguments{
\item{directory}{directory (path/to/files/) where daf and divergence files are stored in your local machine}

\item{test}{which test to perform. Options include: standard (default), DGRP, FWW, asymptotic, iMK}

\item{xlow}{lower limit for asymptotic alpha fit (default=0)}

\item{xhigh}{higher limit for asymptotic alpha fit (default=1)}

\item{fullAnalysis}{decide whether to analyze all files in directory or not (default=TRUE)}

\item{idList}{used when fullAnalysis = F, list of IDs to analyze}
}
\value{
None
}
\description{
Perform any MK test using all files (or a subset of them) in a given directory.
}
\details{
Files in directory must be named: file1*daf*, file1*divergence*, file2*daf*, file2*divergence*, ...
}
\examples{
##example here

}
