% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDaf.R
\docType{data}
\name{myDafData}
\alias{myDafData}
\title{Sample DAF data frame}
\format{A data frame containing polymorphic sites for selected (i) and neutral (0) classes at different DAF categories}
\usage{
myDafData
}
\description{
Data frame containing polymorphism sample data
\itemize{
  \item daf. derived allele frequency (DAF) categories
  \item Pi. number of selected (i) polymorphic sites for each daf category
  \item P0. number of neutral (0) polymorphic sites for each daf category
}
}
\keyword{SampleData}
