% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDaf.R
\docType{data}
\name{mydafdata}
\alias{mydafdata}
\title{Example data frames}
\format{A data frame containing divergent and analyzed sites for selected (i) and neutral (0) classes}
\usage{
mydafdata
}
\description{
Containing basic information to execute the funcionts
\itemize{
  \item daf. derived allele frequency (DAF) categories
  \item Pi. number of selected polymorphic sites for each daf category
  \item P0. number of neutral polymorphic sites for each daf category
}
}
\keyword{datasets}
