% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDivergence.R
\docType{data}
\name{myDivergenceData}
\alias{myDivergenceData}
\title{Sample Divergence data frame}
\format{A data frame containing divergent and analyzed sites for selected (i) and neutral (0) classes}
\usage{
myDivergenceData
}
\description{
Data frame containing divergence sample data
\itemize{
  \item mi. number of selected (i) analyzed sites
  \item Di. number of selected divergent sites
  \item m0. number of neutral (0) analyzed sites
  \item D0. number of neutral divergent sites
}
}
\keyword{SampleData}
