% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDivergence.R
\docType{data}
\name{mydivergencedata}
\alias{mydivergencedata}
\title{Example data frames}
\format{A data frame containing divergent and analyzed sites for selected (i) and neutral (0) classes}
\usage{
mydivergencedata
}
\description{
Data frame containing divergence sample data to execute the funcionts
\itemize{
  \item mi. number of selected analyzed sites
  \item Di. number of selected divergent sites (fixed differences)
  \item m0. number of neutral analyzed sites
  \item D0. number of neutral divergent sites (fixed differences)
}
}
\keyword{datasets}
