% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopFlyAnalysis.R
\name{PopFlyAnalysis}
\alias{PopFlyAnalysis}
\title{PopFlyAnalysis}
\usage{
PopFlyAnalysis(genes = c("gene1", "gene2", "..."), pops = c("pop1", "pop2",
  "..."), recomb = TRUE/FALSE, bins = 0, test = c("standard", "DGRP",
  "FWW", "asymptotic", "iMK"), xlow = 0, xhigh = 1)
}
\arguments{
\item{genes}{list of genes}

\item{pops}{list of populations}

\item{recomb}{group genes according to recombination values (must specify number of bins). TRUE/FALSE. Recomb values (cM/Mb) from Comeron et al. 2012.}

\item{bins}{number of recombination bins to compute (mandatory if recomb = TRUE)}

\item{test}{which test to perform. Options include: standard (default), DGRP, FWW, asymptotic, iMK}

\item{xlow}{lower limit for asymptotic alpha fit (default=0)}

\item{xhigh}{higher limit for asymptotic alpha fit (default=1)}
}
\value{
None
}
\description{
Perform any MK test using a subset of PopFly data defined by custom genes and populations lists
}
\details{
Recombination values (recomb=T) from Comeron et al. 2012 (reference!)
}
\examples{
## List of genes
# mygenes <- c("FBgn0053196", "FBgn0086906", "FBgn0261836", "FBgn0031617", 
#              "FBgn0260965", "FBgn0028899", "FBgn0052580", "FBgn0036181",
#              "FBgn0263077", "FBgn0013733", "FBgn0031857", "FBgn0037836")
## Perform analyses
# PopFlyAnalysis(genes=mygenes , pops=c("RAL","ZI"), recomb=F, test="iMK", xlow=0, xhigh=0.9)
# PopFlyAnalysis(genes=mygenes , pops=c("RAL","ZI"), recomb=T, bins=3, test="DGRP")

}
