% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopFlyAnalysis.R
\name{PopFlyAnalysis}
\alias{PopFlyAnalysis}
\title{PopFlyAnalysis}
\usage{
PopFlyAnalysis(genes = c("gene1", "gene2", "..."), pops = c("pop1", "pop2",
  "..."), recomb = TRUE/FALSE, bins = 0, test = c("standard", "DGRP",
  "FWW", "asymptotic", "iMK"), xlow = 0, xhigh = 1)
}
\arguments{
\item{genes}{list of genes}

\item{pops}{list of populations}

\item{recomb}{group genes according to recombination values (must specify number of bins). TRUE/FALSE}

\item{bins}{number of recombination bins to compute (mandatory if recomb = TRUE)}

\item{test}{which test to perform. Options include: standard (default), DGRP, FWW, iMK, asymptotic, none.}

\item{xlow}{lower limit for asymptotic alpha fit (default=0)}

\item{xhigh}{higher limit for asymptotic alpha fit (default=1)}
}
\value{
None
}
\description{
Perform any MK test using a subset of PopFly data defined by custom genes and populations lists
}
\details{
put details here
}
\examples{
## Load PopFly data into your workspace
# loadPopFly()
## Perform analysis
# mygenes <- c("FBgn0053196", "FBgn0000008")
# PopFlyAnalysis(mygenes , c("RAL","ZI"), recomb=F)
# mygenes <- 
# mygenes <- c("FBgn0000008","FBgn0000014","FBgn0000015","FBgn0000017","FBgn0000018","FBgn0000022",
#              "FBgn0000024","FBgn0000028","FBgn0000032","FBgn0000036","FBgn0000037","FBgn0000038",
#              "FBgn0000039","FBgn0000042","FBgn0000043","FBgn0000044","FBgn0000045","FBgn0000046")
# PopFlyAnalysis(mygenes , c("RAL","ZI"), recomb=T, bins=3, test="asymptotic", xlow=0, xhigh=0.9)

}
