% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopHumanAnalysis.R
\name{PopHumanAnalysis}
\alias{PopHumanAnalysis}
\title{PopHumanAnalysis}
\usage{
PopHumanAnalysis(genes = c("gene1", "gene2", "..."), pops = c("pop1",
  "pop2", "..."), recomb = TRUE/FALSE, bins = 0, test = c("standard",
  "DGRP", "FWW", "asymptotic", "iMK"), xlow = 0, xhigh = 1)
}
\arguments{
\item{genes}{list of genes}

\item{pops}{list of populations}

\item{recomb}{group genes according to recombination values (must specify number of bins). TRUE/FALSE. Recomb values (cM/Mb) from Bheller et al.}

\item{bins}{number of recombination bins to compute (mandatory if recomb = TRUE)}

\item{test}{which test to perform. Options include: standard (default), DGRP, FWW, asymptotic, iMK}

\item{xlow}{lower limit for asymptotic alpha fit (default=0)}

\item{xhigh}{higher limit for asymptotic alpha fit (default=1)}
}
\value{
None
}
\description{
Perform any MK test using a subset of PopHuman data defined by custom genes and populations lists
}
\details{
Recombination values (recomb=T) from Bheller et al. (reference!)
}
\examples{
## List of genes
# mygenes <- c("")
## Perform analyses
# PopHumanAnalysis(genes=mygenes , pops=c(""), recomb=F, test="iMK", xlow=0, xhigh=0.9)
# PopHumanAnalysis(genes=mygenes , pops=c(""), recomb=T, bins=3, test="DGRP")

}
