% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictNLS.R
\name{predictNLS}
\alias{predictNLS}
\title{Get a CI using Monte Carlo simulation based upon a fitted model.  This is necessary because
Directly retrieved from: https://github.com/MesserLab/asymptoticMK
Date = 30/11/2016
Author = REFER AUTHROS? CONTACT WITH PETROV LAB}
\usage{
predictNLS(object, newdata, level = 0.95, nsim = 10000)
}
\arguments{
\item{object}{input object}

\item{newdata}{newdata object}

\item{level}{CI level}

\item{nsim}{nsim montecarlo}
}
\value{
None
}
\description{
Get a CI using Monte Carlo simulation based upon a fitted model.  This is necessary because
Directly retrieved from: https://github.com/MesserLab/asymptoticMK
Date = 30/11/2016
Author = REFER AUTHROS? CONTACT WITH PETROV LAB
}
