% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictNLS.R
\name{predictNLS}
\alias{predictNLS}
\title{predictNLSGet a CI using Monte Carlo simulation based upon a fitted model.  This is necessary because}
\usage{
predictNLS(object, newdata, level = 0.95, nsim = 10000)
}
\arguments{
\item{object}{input object}

\item{newdata}{newdata object}

\item{level}{CI level}

\item{nsim}{nsim montecarlo}
}
\value{
None
}
\description{
getting confidence intervals for non-linear models is a complicated business, apparently. Authors thanks to Andrej-Nikolai Spiess (http://www.dr-spiess.de) for this code.
Directly retrieved from: https://github.com/MesserLab/asymptoticMK
}
\details{
Put details here
}
