% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.R
\name{standard}
\alias{standard}
\title{mkt_standard}
\usage{
standard(daf, divergence)
}
\arguments{
\item{daf}{data frame containing DAF, Pi and P0 values}

\item{divergence}{data frame containing divergent and analyzed sites for selected (i) and neutral (0) classes}
}
\value{
Standard McDonald and Kreitman Test
}
\description{
\code{mkt_standard()} MKT standard formula
}
\details{
The standard McDonald and Kreitman test (MKT) is used to detect the signature of selection at the molecular level. The MKT compares the amount of variation within a species (polymorphism, P) to the divergence (D) between species at two types of sites, one of which is putatively netral and used as the reference to detect selection at the other type of site. In the standard MKT, these sites are synonymous (putatively neutral, 0) and non-synonymous sites (selected sites, i) in a coding region. Under strict neutrality, the ratio of the number of selected and neutral polymorphic sites (Pi/P0) is equal to the ratio of the number of selected and neutral divergence sites (Di/D0).
}
\examples{
standard(mydafdata, mydivergencedata)

}
