% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetPopData.R
\name{subsetPopData}
\alias{subsetPopData}
\title{subsetPopData}
\usage{
subsetPopData(data = c("PopFly", "PopHuman"), genes = c("gene1", "gene2",
  "..."), pops = c("pop1", "pop2", "..."), recomb = TRUE/FALSE, bins = 0,
  xlow, xhigh)
}
\arguments{
\item{data}{input PopFly or PopHuman data}

\item{genes}{list of genes}

\item{pops}{populations to use}

\item{recomb}{group genes according to recombination values (must specify number of bins)}

\item{bins}{number of recombination bins to compute (mandatory if recomb = TRUE)}

\item{xlow}{lower limit for asymptotic alpha fit}

\item{xhigh}{higher limit for asymptotic alpha fit}
}
\value{
None
}
\description{
Perform iMK using a subset of PopFly or PopHuman data defined by custom genes and populations lists
}
\details{
put details here
}
\examples{
## Load PopFly data
# loadPopFly()
## Perform analysis
# mygenes <- c("FBgn0053196", "FBgn0000008")
# subsetPopData("PopFly", mygenes , c("RAL","ZI"), recomb=F, bins=0, xlow=0, xhigh=0.9)

}
