% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetPopData.R
\name{subsetPopData}
\alias{subsetPopData}
\title{subsetPopData}
\usage{
subsetPopData(data = c("PopFly", "PopHuman"), genes = c("gene1", "gene2",
  "..."), pops = c("pop1", "pop2", "..."), recomb = TRUE/FALSE, bins = 0)
}
\arguments{
\item{data}{input PopFly or PopHuman data}

\item{genes}{Drosophila gene list}

\item{pops}{Drosophila populations}

\item{recomb}{group genes according to recombination values (must specify number of bins)}

\item{bins}{number of recombination bins to compute (mandatory if recomb = TRUE)}
}
\value{
None
}
\description{
Perform iMK using a subset of PopFly or PopHuman data defined by custom genes and populations lists
}
\details{
put details here
}
\examples{
## Load PopFly data
# loadPopFly()
## Perform analysis
# subsetPopData("PopFly", c("FBgn0053196", "FBgn0000008"),c("RAL","ZI"), F)

}
