% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popfly.R
\name{subsetPopfly}
\alias{subsetPopfly}
\title{Extract a list of genes from PopFly
Date = 30/11/2016
Author = Jesús Murga, Marta Coronado}
\usage{
subsetPopfly(datos = "data input", genes = c("gene1"),
  POP = c("population"), recomb = TRUE/FALSE, bin)
}
\arguments{
\item{datos}{PopFly dataset preload in package}

\item{genes}{Drosophila gene list}

\item{POP}{Drosophila gene population}

\item{recomb}{Retrieve genes by recombinations values}

\item{bin}{recombination value}
}
\value{
None
}
\description{
Extract a list of genes from PopFly
Date = 30/11/2016
Author = Jesús Murga, Marta Coronado
}
