% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popfly.R
\name{subsetPopFly}
\alias{subsetPopFly}
\title{Extract a list of genes from PopFly
Date = 30/11/2016
Author = Jesús Murga, Marta Coronado, Sergi Hervas}
\usage{
subsetPopFly(data = "data", genes = c("gene1", "gene2", "..."),
  pops = c("pop1", "pop2", "..."), recomb = TRUE/FALSE, bins)
}
\arguments{
\item{genes}{Drosophila gene list}

\item{pops}{Drosophila populations}

\item{recomb}{group genes according to recombination values (must specify bin)}

\item{bin}{number of recombination bins to compute}
}
\value{
None
}
\description{
Extract a list of genes from PopFly
Date = 30/11/2016
Author = Jesús Murga, Marta Coronado, Sergi Hervas
}
