% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popfly.R
\name{subsetPopFly}
\alias{subsetPopFly}
\title{subsetPopfly}
\usage{
subsetPopFly(genes = c("gene1", "gene2", "..."), pops = c("pop1", "pop2",
  "..."), recomb = TRUE/FALSE, bins)
}
\arguments{
\item{genes}{Drosophila gene list}

\item{pops}{Drosophila populations}

\item{recomb}{group genes according to recombination values (must specify number of bins)}

\item{bins}{number of recombination bins to compute (mandatory if recomb = TRUE)}
}
\value{
None
}
\description{
Extract a list of genes from PopFly
}
\details{
put details here
}
\examples{

}
