% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMK.R, R/theme_Publication.R
\name{theme_Publication}
\alias{theme_Publication}
\alias{theme_Publication}
\title{iMK compute all the MKT extensions 
Date = 30/11/2016
Author = Sergi Hervás, Marta Coronado}
\usage{
theme_Publication(base_size = 14, base_family = "helvetica")

theme_Publication(base_size = 14, base_family = "helvetica")
}
\arguments{
\item{base_size}{base size required from theme_Publication}

\item{base_family}{font to load in theme_Publication}

\item{daf}{dad file}

\item{divergence}{div file}

\item{xlow}{fit curve}

\item{xhigh}{curv}
}
\value{
None

plot theme
}
\description{
EXPLANATION

Date = 04/07/2015
Author = Koundinya Desiraju
}
\details{
Default theme used for plot images. From http://rpubs.com/Koundy/71792
}
