% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{ggplot theme for publication ready Plots}
\usage{
theme_Publication(base_size = 14, base_family = "sans")
}
\arguments{
\item{base_size}{base size required from theme_Publication}

\item{base_family}{font to load in theme_Publication}
}
\value{
plot theme
}
\description{
Date = 04/07/2015
Author = Koundinya Desiraju
}
\details{
Default theme used for plot images. From http://rpubs.com/Koundy/71792
}
\examples{
# theme_Publication(14, "sans")
}
