% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themePublication.R
\name{themePublication}
\alias{themePublication}
\title{ggplot Theme for publication ready plots}
\usage{
themePublication(base_size = 14, base_family = "sans")
}
\arguments{
\item{base_size}{base size required from themePublication}

\item{base_family}{font to load in themePublication}
}
\value{
plot theme
}
\description{
Theme with the configuration and parameters necessary to generate publication ready plots using ggplot
}
\details{
Theme used for plot images developed by Koundinya Desiraju (04/07/2015). Code adapted from http://rpubs.com/Koundy/71792.
}
